feat: move most of chip saturation to graph on chip
