+++ src/data/GraphOnChip.ts
@@ -1094,6 +1094,34 @@ export class ComputeNode {
             }
         });
 
+        let offChipLinkIds: string[] = [];
+
+        switch (node.type) {
+            case ComputeNodeType.DRAM:
+                offChipLinkIds =
+                    node.dramChannel?.links.map((l) => {
+                        return l.uid;
+                    }) || [];
+                break;
+            case ComputeNodeType.ETHERNET:
+                offChipLinkIds =
+                    [...node.internalLinks.values()].map((l) => {
+                        return l.uid;
+                    }) || [];
+                break;
+
+            case ComputeNodeType.PCIE:
+                offChipLinkIds =
+                    [...node.internalLinks].map(([l]) => {
+                        return l.uid;
+                    }) || [];
+                break;
+            default:
+                break;
+        }
+
+        node.offchipLinkIds = offChipLinkIds;
+
         // Associate with operation
         const opName: OperationName = nodeJSON.op_name;
         if (opName) {
@@ -1131,6 +1159,8 @@ export class ComputeNode {
     public links: Map<any, NOCLink> = new Map();
 
     public noc0and1Links: NOCLink[] = [];
+
+    public offchipLinkIds: string[] = [];
 
     /** @description Off chip links that are not part of the NOC, excluding DRAM links */
     public internalLinks: Map<any, NetworkLink> = new Map();
