+++ src/renderer/components/node-grid-elements-components/OffChipNodeLinkCongestionLayer.tsx
@@ -5,7 +5,6 @@
 import { FC, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { ComputeNode } from '../../../data/GraphOnChip';
-import { ComputeNodeType } from '../../../data/Types';
 import { getLinkSaturation } from '../../../data/store/selectors/linkSaturation.selectors';
 import { calculateLinkCongestionColor, getOffChipCongestionStyles, toRGBA } from '../../../utils/DrawingAPI';
 import type { LinkState } from '../../../data/StateTypes';
@@ -29,33 +28,7 @@ const OffChipNodeLinkCongestionLayer: FC<OffChipNodeLinkCongestionLayerProps> = 
     const linkSaturationTreshold = useSelector(getLinkSaturation);
 
     const saturation = useMemo(() => {
-        let offChipLinkIds: string[] = [];
-
-        switch (node.type) {
-            case ComputeNodeType.DRAM:
-                offChipLinkIds =
-                    node.dramChannel?.links.map((link) => {
-                        return link.uid;
-                    }) || [];
-                break;
-            case ComputeNodeType.ETHERNET:
-                offChipLinkIds =
-                    [...node.internalLinks.values()].map((link) => {
-                        return link.uid;
-                    }) || [];
-                break;
-
-            case ComputeNodeType.PCIE:
-                offChipLinkIds =
-                    [...node.internalLinks].map(([link]) => {
-                        return link.uid;
-                    }) || [];
-                break;
-            default:
-                break;
-        }
-
-        const saturationValues = offChipLinkIds.map((linkId) => linksData[linkId]?.saturation) || [0];
+        const saturationValues = node.offchipLinkIds.map((linkId) => linksData[linkId]?.saturation) || [0];
 
         return Math.max(...saturationValues) || 0;
     }, [linksData, node]);
